+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 ### App Center Crashes
 
 * **[Fix]** Remove the multiple attachments warning as that is now supported by the portal.
+* **[Fix]** Change minidump filter to use file extension instead of name.
+* **[Fix]** Fix removing minidump files when the sending crash report was discarded.
 
 ### App Center Distribute